Use new timetable PDF generation in static sites
